refactor(form field container): rename component
