--- packages/react-magma-dom/src/components/FormField/index.ts
@@ -1,1 +1,0 @@
-export * from './FormField';
