+++ packages/react-magma-dom/src/components/FormFieldContainer/FormFieldContainer.test.js
@@ -1,14 +1,16 @@
 import React from 'react';
 import { axe } from 'jest-axe';
-import { FormField } from '.';
+import { FormFieldContainer } from '.';
 import { render } from '@testing-library/react';
 import { I18nContext, defaultI18n } from '../../i18n';
 
 const TEXT = 'Test Text';
 
-describe('FormField', () => {
+describe('FormFieldContainer', () => {
   it('should render the visually hidden component', () => {
-    const { container, getByText } = render(<FormField>{TEXT}</FormField>);
+    const { container, getByText } = render(
+      <FormFieldContainer>{TEXT}</FormFieldContainer>
+    );
 
     expect(getByText(TEXT)).toBeInTheDocument();
   });
@@ -16,14 +18,16 @@ describe('FormField', () => {
   it('should find element by testId', () => {
     const testId = 'test-id';
     const { getByTestId } = render(
-      <FormField testId={testId}>{TEXT}</FormField>
+      <FormFieldContainer testId={testId}>{TEXT}</FormFieldContainer>
     );
 
     expect(getByTestId(testId)).toBeInTheDocument();
   });
 
   it('Does not violate accessibility standards', () => {
-    const { container } = render(<FormField>{TEXT}</FormField>);
+    const { container } = render(
+      <FormFieldContainer>{TEXT}</FormFieldContainer>
+    );
 
     return axe(container.innerHTML).then(result => {
       return expect(result).toHaveNoViolations();
