+++ packages/react-magma-dom/src/components/TimePicker/useTimePicker.ts
@@ -1,18 +1,17 @@
 import * as React from 'react';
 import { useGenerateId } from '../../utils';
-import { FormFieldBaseProps } from '../FormField';
+import { FormFieldContainerBaseProps } from '../FormFieldContainer';
 
 import { enUS } from 'date-fns/locale';
 import { I18nContext } from '../../i18n';
 
-export interface UseTimePickerProps extends FormFieldBaseProps {
+export interface UseTimePickerProps extends FormFieldContainerBaseProps {
   /**
    * Style properties for the outer container
    */
   containerStyle?: React.CSSProperties;
   /**
-   * This is internal to stay consistent with our props tables and our messaging around spreading default props to elements.
-   * @internal
+   * ID of the hidden input that stores the time value.  Also the prefix for other fields.
    */
   id?: string;
   /**
