+++ packages/react-magma-dom/src/index.ts
@@ -222,4 +222,4 @@ export * from './components/Datagrid';
 export * from './components/Transition';
 export * from './components/Drawer';
 
-export * from './components/FormField';
+export * from './components/FormFieldContainer';
